refactor: Refactored renumbering logic into ToolsPlugin
The renumbering logic previously present in PasteAndRenumberAction has been moved to the ToolsPlugin class. This change improves code organization and allows for easier reuse of the renumbering functionality. Additionally, a new DuplicateAndRenumberAction was added to the edit menu, extending the application's capabilities.
