+++ lexxpluss_tools/src/org/openstreetmap/josm/plugins/lexxpluss/PasteAndRenumberAction.java
@@ -8,13 +8,10 @@ package org.openstreetmap.josm.plugins.lexxpluss;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
-import java.util.Arrays;
 import static org.openstreetmap.josm.gui.help.HelpUtil.ht;
 import static org.openstreetmap.josm.tools.I18n.tr;
 import org.openstreetmap.josm.actions.AbstractPasteAction;
 import org.openstreetmap.josm.gui.MainApplication;
-import org.openstreetmap.josm.data.osm.Node;
-import org.openstreetmap.josm.data.osm.Way;
 import org.openstreetmap.josm.tools.Shortcut;
 
 /**
@@ -38,46 +35,6 @@ public class PasteAndRenumberAction extends AbstractPasteAction {
         super.actionPerformed(e);
         var ds = getLayerManager().getEditDataSet();
         var objs = ds.getSelectedNodesAndWays();
-        if (objs != null) {
-            objs.forEach(o -> {
-                if (o instanceof Node)
-                    nodeIdRenumber((Node)o);
-                else if (o instanceof Way)
-                    wayIdRenumber((Way)o);
-            });
-        }
-    }
-
-    /**
-     * Renumber node tags.
-     * @param node the node
-     */
-    private void nodeIdRenumber(Node node) {
-        Arrays.asList("agv_node_id", "intermediate_goal_id")
-                .forEach(key -> {
-                    if (node.hasKey(key)) {
-                        var max = ToolsPlugin.getMaxId(node.getDataSet().getNodes(), key);
-                        node.put(key, Integer.toString(max + 1));
-                    }
-                });
-    }
-
-    /**
-     * Renumber way tags.
-     * @param way the way
-     */
-    private void wayIdRenumber(Way way) {
-        Arrays.asList("goal_id", "space_id")
-                .forEach(key -> {
-                    if (way.hasKey(key)) {
-                        var max = ToolsPlugin.getMaxId(way.getDataSet().getWays(), key);
-                        var s = Integer.toString(max + 1);
-                        way.put(key, s);
-                        if (key.equals("space_id")) {
-                            way.put("area_name" , "park" + s);
-                        }
-                    }
-                });
-        way.getNodes().forEach(n -> nodeIdRenumber(n));
+        ToolsPlugin.renumber(objs);
     }
 }
